+++ src/Client.ts
@@ -1,7 +1,6 @@
 import { io as SocketClient, Socket } from "socket.io-client";
 import { ClientMessages, ServerMessages } from "./API";
 import ConnectionMonitor from "./ConnectionMonitor";
-import { ConnectionState } from "./RoomClient";
 
 /**
  * Client is a typed wrapper for raw socket events sent to and from the server.
@@ -13,29 +12,16 @@ export default class Client {
   /**
    * connect is a Client constructor that will wait until it is connected
    */
-  static async connect(
-    url: string,
-    broadcastConnectionState: boolean
-  ): Promise<Client> {
-    const client = new this(url, broadcastConnectionState);
+  static async connect(url: string): Promise<Client> {
+    const client = new this(url);
     await client.waitFor("connect");
     return client;
   }
 
-  protected constructor(
-    url: string,
-    private broadcastConnectionState?: boolean
-  ) {
+  protected constructor(url: string) {
     this.socket = SocketClient(url, { transports: ["websocket"] });
     this.connectionMonitor = new ConnectionMonitor(this.socket, 500);
-    if (this.broadcastConnectionState) {
-      this.connectionMonitor.emitter.on("quality", this.emitConnectionState);
-    }
   }
-
-  private emitConnectionState = (currentQuality: ConnectionState) => {
-    this.emit("connectionState", currentQuality);
-  };
 
   on<E extends keyof ServerMessages>(
     name: E,
@@ -52,9 +38,6 @@ export default class Client {
   }
 
   close(): void {
-    if (this.broadcastConnectionState) {
-      this.connectionMonitor.emitter.off("quality", this.emitConnectionState);
-    }
     this.socket.close();
   }